Install repose and use Coco pose type directly
