+++ Pipfile
@@ -6,6 +6,7 @@ verify_ssl = true
 [packages]
 fire = "*"
 websockets = "*"
+repose = {editable = true, path = "/Users/how/Own/repose"}
 
 [dev-packages]
 black = "==20.8b1"
@@ -17,10 +18,6 @@ pre-commit = "*"
 
 [requires]
 python_version = "3.9"
-
-[isort]
-line_length = 80
-use_parentheses = true
 
 [scripts]
 cli = "python -m src.services.cli"
