+++ setup.cfg
@@ -1,0 +1,9 @@
+[isort]
+use_parentheses=True
+line_length=80
+
+
+[mypy]
+
+[mypy-repose.*]
+ignore_missing_imports = true
